Move device_data_references and a few traits to new file.
