+++ cpp/include/cudf/ast/detail/expression_evaluator.cuh
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include <cudf/ast/detail/expression_parser.hpp>
+#include <cudf/ast/detail/device_data_references.hpp>
 #include <cudf/ast/detail/operators.hpp>
 #include <cudf/ast/expressions.hpp>
 #include <cudf/column/column_device_view.cuh>
